fit-hero-ai-service/requirements.txt
crewai==0.41.1
crewai-tools==0.4.26
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.6.1
python-dotenv==1.0.0
google-generativeai==0.8.3
litellm==1.52.0
requests==2.31.0
python-multipart==0.0.6
